+++ tests/compiler/std/array.untouched.wat
@@ -7646,7 +7646,7 @@
   local.get $1
   call $~lib/array/Array<f32>#sort
  )
- (func $~lib/builtins/isNaN<f32> (; 140 ;) (type $FUNCSIG$if) (param $0 f32) (result i32)
+ (func $~lib/number/isNaN<f32> (; 140 ;) (type $FUNCSIG$if) (param $0 f32) (result i32)
   local.get $0
   local.get $0
   f32.ne
@@ -7707,11 +7707,11 @@
      local.get $0
      local.get $3
      call $~lib/array/Array<f32>#__get
-     call $~lib/builtins/isNaN<f32>
+     call $~lib/number/isNaN<f32>
      local.get $1
      local.get $3
      call $~lib/array/Array<f32>#__get
-     call $~lib/builtins/isNaN<f32>
+     call $~lib/number/isNaN<f32>
      i32.eq
      if
       br $continue|0
@@ -8301,7 +8301,7 @@
   local.get $1
   call $~lib/array/Array<f64>#__unchecked_get
  )
- (func $~lib/builtins/isNaN<f64> (; 150 ;) (type $FUNCSIG$id) (param $0 f64) (result i32)
+ (func $~lib/number/isNaN<f64> (; 150 ;) (type $FUNCSIG$id) (param $0 f64) (result i32)
   local.get $0
   local.get $0
   f64.ne
@@ -8362,11 +8362,11 @@
      local.get $0
      local.get $3
      call $~lib/array/Array<f64>#__get
-     call $~lib/builtins/isNaN<f64>
+     call $~lib/number/isNaN<f64>
      local.get $1
      local.get $3
      call $~lib/array/Array<f64>#__get
-     call $~lib/builtins/isNaN<f64>
+     call $~lib/number/isNaN<f64>
      i32.eq
      if
       br $continue|0
@@ -11392,7 +11392,7 @@
   if
    i32.const 0
    i32.const 4360
-   i32.const 42
+   i32.const 40
    i32.const 4
    call $~lib/builtins/abort
    unreachable
@@ -12165,7 +12165,7 @@
   if
    i32.const 0
    i32.const 4360
-   i32.const 198
+   i32.const 196
    i32.const 4
    call $~lib/builtins/abort
    unreachable
@@ -13215,7 +13215,7 @@
   local.get $2
   return
  )
- (func $~lib/builtins/isFinite<f64> (; 245 ;) (type $FUNCSIG$id) (param $0 f64) (result i32)
+ (func $~lib/number/isFinite<f64> (; 245 ;) (type $FUNCSIG$id) (param $0 f64) (result i32)
   local.get $0
   local.get $0
   f64.sub
@@ -14577,11 +14577,11 @@
    return
   end
   local.get $0
-  call $~lib/builtins/isFinite<f64>
+  call $~lib/number/isFinite<f64>
   i32.eqz
   if
    local.get $0
-   call $~lib/builtins/isNaN<f64>
+   call $~lib/number/isNaN<f64>
    if
     i32.const 5392
     call $~lib/rt/pure/__retain
@@ -14649,11 +14649,11 @@
    return
   end
   local.get $2
-  call $~lib/builtins/isFinite<f64>
+  call $~lib/number/isFinite<f64>
   i32.eqz
   if
    local.get $2
-   call $~lib/builtins/isNaN<f64>
+   call $~lib/number/isNaN<f64>
    if
     local.get $0
     i32.const 78
